pyproject.toml
[tool.poetry]
name = "marathonic"
version = "0.1.0"
description = ""
authors = ["Blake Jeong <jmchan2432@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2"
django-debug-toolbar = "^5.2.0"
django-allauth = {extras = ["socialaccount"], version = "^65.8.0"}
django-environ = "^0.12.0"
django-extensions = "^4.1"
pillow = "^11.2.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
pandas = "^2.2.3"
django-filter = "^25.1"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
